feat: add single file, directory, or dynamic output options
Allows for `output` option in params passed to node-sass-extra; if a single file path is given, all
sources are concatenated and written to disk. If a directory is given, files are compiled and
flattened. If a function/dynamic output is given, it will execute on the source path and use the
returned path

closes #6