feat: support for `outFile` and `sourceMap` node-sass config options
`outFile` supports the same features as `output` but will not write to disk. Source maps will be
written to disk when `sourceMap` is provided alongside `output`. If a function is passed
to `sourceMap`, it will execute on the output path and use the returned path. Logic used to
determine `output` was moved out of `render` and into a stand-alone utility. Contatenation of
multiple sources to a single output was removed temporarily due to incompatibility with source maps.
Plus minor updates to utility functions and comments.

closes #18