refactor: `getCompilerTasks` broken out into multiple functions
To improve readability, `getCompilerTasks` was broken out into three logical, stand-alone functions.
As a result, the implementation is simpler and more maintainable. Also, by happenstance, the dynamic
iterators for `output` and `outFile` now support the return of directories and `sourceMap` also now
supports directories. `isFile` was also refactored to test if a given path is a file more generally,
instead of checking for specific extensions.

closes #21